llm_engineering/experimentation/python/requirements.txt
# Requirements for ollama.py
# Install with: pip install -r requirements.txt

python-dotenv>=1.0.0
openai>=1.0.0
gradio>=4.0.0
